requirements.txt
openwakeword
pywhispercpp
sounddevice
numpy
requests
psutil
noisereduce
scipy
webrtcvad-wheels
